+++ src/birth/game-play-initializer.cpp
@@ -16,6 +16,7 @@
 #include "player/player-spell-status.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon-info.h"
 #include "system/floor-type-definition.h"
