+++ src/birth/inventory-initializer.cpp
@@ -30,6 +30,7 @@
 #include "sv-definition/sv-wand-types.h"
 #include "sv-definition/sv-weapon-types.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/enum-converter.h"
