+++ src/core/visuals-reseter.cpp
@@ -2,6 +2,7 @@
 #include "game-option/special-options.h"
 #include "io/read-pref-file.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/monster-race-info.h"
 #include "system/player-type-definition.h"
 #include "system/terrain-type-definition.h"
