+++ src/flavor/tval-description-switcher.cpp
@@ -11,6 +11,7 @@
 #include "object-enchant/trg-types.h"
 #include "object/tval-types.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/item-entity.h"
 #include "system/monster-race-info.h"
 #include "util/bit-flags-calculator.h"
