+++ src/info-reader/baseitem-reader.cpp
@@ -15,6 +15,7 @@
 #include "object-enchant/tr-types.h"
 #include "object/tval-types.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "term/gameterm.h"
 #include "util/bit-flags-calculator.h"
 #include "util/enum-converter.h"
