+++ src/io/interpret-pref-file.cpp
@@ -14,6 +14,7 @@
 #include "io/input-key-requester.h"
 #include "io/tokenizer.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/enums/game-option-page.h"
 #include "system/monster-race-info.h"
 #include "system/player-type-definition.h"
