+++ src/item-info/flavor-initializer.cpp
@@ -8,6 +8,7 @@
 #include "object/tval-types.h"
 #include "system/angband-system.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 
 /*!
  * @brief ゲーム開始時に行われるベースアイテムの初期化ルーチン
