+++ src/knowledge/knowledge-items.cpp
@@ -20,6 +20,7 @@
 #include "perception/object-perception.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
