+++ src/load/item/item-loader-base.cpp
@@ -4,6 +4,7 @@
 #include "load/load-util.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "util/bit-flags-calculator.h"
 #include "util/enum-converter.h"
