+++ src/load/old/load-v1-5-0.cpp
@@ -36,6 +36,7 @@
 #include "system/angband-exceptions.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/dungeon-info.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor-type-definition.h"
