+++ src/main/game-data-initializer.cpp
@@ -13,6 +13,7 @@
 #include "game-option/option-types-table.h"
 #include "system/alloc-entries.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/dungeon-info.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
