+++ src/market/building-craft-fix.cpp
@@ -20,6 +20,7 @@
 #include "sv-definition/sv-other-types.h"
 #include "sv-definition/sv-weapon-types.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
