+++ src/monster-floor/monster-death.cpp
@@ -31,6 +31,7 @@
 #include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon-info.h"
 #include "system/enums/monrace/monrace-id.h"
