+++ src/store/rumor.cpp
@@ -8,6 +8,7 @@
 #include "system/angband-exceptions.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/dungeon-info.h"
 #include "system/item-entity.h"
 #include "system/monster-race-info.h"
