+++ src/system/artifact-type-definition.cpp
@@ -2,6 +2,7 @@
 #include "artifact/fixed-art-types.h"
 #include "object/tval-types.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/item-entity.h"
 
 ArtifactType::ArtifactType()
