+++ src/system/floor-type-definition.cpp
@@ -6,6 +6,7 @@
 #include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/dungeon-info.h"
 #include "system/enums/grid-count-kind.h"
 #include "system/gamevalue.h"
