+++ src/wizard/items-spoiler.cpp
@@ -7,6 +7,7 @@
 #include "object/object-value.h"
 #include "system/angband-system.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "term/z-form.h"
