+++ src/wizard/wizard-item-modifier.cpp
@@ -28,6 +28,7 @@
 #include "system/alloc-entries.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-list.h"
 #include "system/floor-type-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
